Split kafka connection
