+++ examples/main.tf
@@ -79,9 +79,8 @@ resource "materialize_connection" "privatelink_conn" {
   aws_privatelink_service_name       = "com.amazonaws.us-east-1.materialize.example"
   aws_privatelink_availability_zones = ["use1-az2", "use1-az1"]
 }
-resource "materialize_connection" "example_kafka_privatelink_conn" {
+resource "materialize_connection_kafka" "example_kafka_privatelink_conn" {
   name            = "example_kafka_privatelink_conn"
-  connection_type = "KAFKA"
   kafka_broker {
     broker                 = "b-1.hostname-1:9096"
     target_group_port      = "9001"
@@ -109,9 +108,8 @@ resource "materialize_connection" "example_postgres_connection" {
 }
 
 # Create a Kafka Connection
-resource "materialize_connection" "example_kafka_connection" {
+resource "materialize_connection_kafka" "example_kafka_connection" {
   name            = "example_kafka_connection"
-  connection_type = "KAFKA"
   kafka_broker {
     broker = "b-1.hostname-1:9096"
   }
