+++ examples/resources/materialize_connection_kafka/import.sh
@@ -1,0 +1,2 @@
+#Connections can be imported using the connection id:
+terraform import materialize_connection_kafka.example <connection_id>
