+++ materialize/provider.go
@@ -59,6 +59,7 @@ func Provider() *schema.Provider {
 			"materialize_cluster":               resources.Cluster(),
 			"materialize_cluster_replica":       resources.ClusterReplica(),
 			"materialize_connection":            resources.Connection(),
+			"materialize_connection_kafka":      resources.ConnectionKafka(),
 			"materialize_database":              resources.Database(),
 			"materialize_schema":                resources.Schema(),
 			"materialize_secret":                resources.Secret(),
