+++ materialize/resources/resource_connection_test.go
@@ -22,170 +22,6 @@ func TestResourceConnectoinReadId(t *testing.T) {
 	`, b.ReadId())
 }
 
-func TestResourceConnectionCreateSsh(t *testing.T) {
-	r := require.New(t)
-
-	b := newConnectionBuilder("ssh_conn", "schema", "database")
-	b.ConnectionType("SSH TUNNEL")
-	b.SSHHost("localhost")
-	b.SSHPort(123)
-	b.SSHUser("user")
-	r.Equal(`CREATE CONNECTION database.schema.ssh_conn TO SSH TUNNEL (HOST 'localhost', USER 'user', PORT 123);`, b.Create())
-
-}
-
-func TestResourceConnectionCreateAwsPrivateLink(t *testing.T) {
-	r := require.New(t)
-
-	b := newConnectionBuilder("privatelink_conn", "schema", "database")
-	b.ConnectionType("AWS PRIVATELINK")
-	b.PrivateLinkServiceName("com.amazonaws.us-east-1.materialize.example")
-	b.PrivateLinkAvailabilityZones([]string{"use1-az1", "use1-az2"})
-	r.Equal(`CREATE CONNECTION database.schema.privatelink_conn TO AWS PRIVATELINK (SERVICE NAME 'com.amazonaws.us-east-1.materialize.example',AVAILABILITY ZONES ('use1-az1', 'use1-az2'));`, b.Create())
-}
-
-func TestResourceConnectionCreateKafka(t *testing.T) {
-	r := require.New(t)
-	b := newConnectionBuilder("kafka_conn", "schema", "database")
-	b.ConnectionType("KAFKA")
-	b.KafkaBrokers([]KafkaBroker{
-		{
-			Broker: "localhost:9092",
-		},
-	})
-	b.KafkaProgressTopic("topic")
-	b.KafkaSASLMechanisms("PLAIN")
-	b.KafkaSASLUsername("user")
-	b.KafkaSASLPassword("password")
-	r.Equal(`CREATE CONNECTION database.schema.kafka_conn TO KAFKA (BROKERS ('localhost:9092'), PROGRESS TOPIC 'topic', SASL MECHANISMS = 'PLAIN', SASL USERNAME = 'user', SASL PASSWORD = SECRET password);`, b.Create())
-}
-
-func TestResourceConnectionCreateKafkaMultipleBrokers(t *testing.T) {
-	r := require.New(t)
-	b := newConnectionBuilder("kafka_conn", "schema", "database")
-	b.ConnectionType("KAFKA")
-	b.KafkaBrokers([]KafkaBroker{
-		{
-			Broker: "localhost:9092",
-		},
-		{
-			Broker: "localhost:9093",
-		},
-	})
-	b.KafkaProgressTopic("topic")
-	b.KafkaSASLMechanisms("PLAIN")
-	b.KafkaSASLUsername("user")
-	b.KafkaSASLPassword("password")
-	r.Equal(`CREATE CONNECTION database.schema.kafka_conn TO KAFKA (BROKERS ('localhost:9092', 'localhost:9093'), PROGRESS TOPIC 'topic', SASL MECHANISMS = 'PLAIN', SASL USERNAME = 'user', SASL PASSWORD = SECRET password);`, b.Create())
-}
-
-func TestResourceConnectionCreateKafkaSsh(t *testing.T) {
-	r := require.New(t)
-	b := newConnectionBuilder("kafka_conn", "schema", "database")
-	b.ConnectionType("KAFKA")
-	b.KafkaBrokers([]KafkaBroker{
-		{
-			Broker: "localhost:9092",
-		},
-	})
-	b.KafkaProgressTopic("topic")
-	b.KafkaSASLMechanisms("PLAIN")
-	b.KafkaSASLUsername("user")
-	b.KafkaSASLPassword("password")
-	b.KafkaSSHTunnel("ssh_conn")
-	r.Equal(`CREATE CONNECTION database.schema.kafka_conn TO KAFKA (BROKERS ('localhost:9092' USING SSH TUNNEL ssh_conn), PROGRESS TOPIC 'topic', SASL MECHANISMS = 'PLAIN', SASL USERNAME = 'user', SASL PASSWORD = SECRET password);`, b.Create())
-}
-
-func TestResourceConnectionCreateKafkaBrokers(t *testing.T) {
-	r := require.New(t)
-	b := newConnectionBuilder("kafka_conn", "schema", "database")
-	b.ConnectionType("KAFKA")
-	b.KafkaBrokers([]KafkaBroker{
-		{
-			Broker: "localhost:9092",
-		},
-		{
-			Broker: "localhost:9093",
-		},
-	})
-	b.KafkaProgressTopic("topic")
-	b.KafkaSASLMechanisms("PLAIN")
-	b.KafkaSASLUsername("user")
-	b.KafkaSASLPassword("password")
-	r.Equal(`CREATE CONNECTION database.schema.kafka_conn TO KAFKA (BROKERS ('localhost:9092', 'localhost:9093'), PROGRESS TOPIC 'topic', SASL MECHANISMS = 'PLAIN', SASL USERNAME = 'user', SASL PASSWORD = SECRET password);`, b.Create())
-}
-
-func TestResourceConnectionCreateKafkaBrokersSsh(t *testing.T) {
-	r := require.New(t)
-	b := newConnectionBuilder("kafka_conn", "schema", "database")
-	b.ConnectionType("KAFKA")
-	b.KafkaBrokers([]KafkaBroker{
-		{
-			Broker: "localhost:9092",
-		},
-		{
-			Broker: "localhost:9093",
-		},
-	})
-	b.KafkaProgressTopic("topic")
-	b.KafkaSASLMechanisms("PLAIN")
-	b.KafkaSASLUsername("user")
-	b.KafkaSASLPassword("password")
-	b.KafkaSSHTunnel("ssh_conn")
-	r.Equal(`CREATE CONNECTION database.schema.kafka_conn TO KAFKA (BROKERS ('localhost:9092' USING SSH TUNNEL ssh_conn,'localhost:9093' USING SSH TUNNEL ssh_conn), PROGRESS TOPIC 'topic', SASL MECHANISMS = 'PLAIN', SASL USERNAME = 'user', SASL PASSWORD = SECRET password);`, b.Create())
-}
-
-func TestResourceConnectionCreateKafkaSsl(t *testing.T) {
-	r := require.New(t)
-	b := newConnectionBuilder("kafka_conn", "schema", "database")
-	b.ConnectionType("KAFKA")
-	b.KafkaBrokers([]KafkaBroker{
-		{
-			Broker: "localhost:9092",
-		},
-	})
-	b.KafkaProgressTopic("topic")
-	b.KafkaSSLKey("key")
-	b.KafkaSSLCert("cert")
-	b.KafkaSSLCa("ca")
-	r.Equal(`CREATE CONNECTION database.schema.kafka_conn TO KAFKA (BROKERS ('localhost:9092'), PROGRESS TOPIC 'topic', SSL CERTIFICATE AUTHORITY = SECRET ca, SSL CERTIFICATE = SECRET cert, SSL KEY = SECRET key);`, b.Create())
-}
-
-func TestResourceConnectionCreateConfluentSchemaRegistry(t *testing.T) {
-	r := require.New(t)
-	b := newConnectionBuilder("csr_conn", "schema", "database")
-	b.ConnectionType("CONFLUENT SCHEMA REGISTRY")
-	b.ConfluentSchemaRegistryUrl("http://localhost:8081")
-	b.ConfluentSchemaRegistryUsername("user")
-	b.ConfluentSchemaRegistryPassword("password")
-	r.Equal(`CREATE CONNECTION database.schema.csr_conn TO CONFLUENT SCHEMA REGISTRY (URL 'http://localhost:8081', USERNAME = 'user', PASSWORD = SECRET password);`, b.Create())
-
-}
-
-func TestResourceConnectionKafkaAwsPrivatelink(t *testing.T) {
-	r := require.New(t)
-	b := newConnectionBuilder("kafka_conn", "schema", "database")
-	b.ConnectionType("KAFKA")
-	b.KafkaBrokers([]KafkaBroker{
-		{
-			Broker:                "b-1.hostname-1:9096",
-			TargetGroupPort:       9001,
-			AvailabilityZone:      "use1-az1",
-			PrivateLinkConnection: "privatelink_conn",
-		},
-		{
-			Broker:                "b-1.hostname-1:9097",
-			TargetGroupPort:       9002,
-			AvailabilityZone:      "use1-az2",
-			PrivateLinkConnection: "privatelink_conn",
-		},
-	})
-	b.KafkaSASLMechanisms("PLAIN")
-	b.KafkaSASLUsername("user")
-	b.KafkaSASLPassword("password")
-	r.Equal(`CREATE CONNECTION database.schema.kafka_conn TO KAFKA (BROKERS ('b-1.hostname-1:9096' USING AWS PRIVATELINK privatelink_conn (PORT 9001, AVAILABILITY ZONE 'use1-az1'), 'b-1.hostname-1:9097' USING AWS PRIVATELINK privatelink_conn (PORT 9002, AVAILABILITY ZONE 'use1-az2')), SASL MECHANISMS = 'PLAIN', SASL USERNAME = 'user', SASL PASSWORD = SECRET password);`, b.Create())
-}
-
 func TestResourceConnectionRename(t *testing.T) {
 	r := require.New(t)
 	b := newConnectionBuilder("connection", "schema", "database")
@@ -213,4 +49,38 @@ func TestResourceConnectionReadParams(t *testing.T) {
 		JOIN mz_databases
 			ON mz_schemas.database_id = mz_databases.id
 		WHERE mz_connections.id = 'u1';`, b)
+}
+
+// here are common ^
+func TestResourceConnectionCreateSsh(t *testing.T) {
+	r := require.New(t)
+
+	b := newConnectionBuilder("ssh_conn", "schema", "database")
+	b.ConnectionType("SSH TUNNEL")
+	b.SSHHost("localhost")
+	b.SSHPort(123)
+	b.SSHUser("user")
+	r.Equal(`CREATE CONNECTION database.schema.ssh_conn TO SSH TUNNEL (HOST 'localhost', USER 'user', PORT 123);`, b.Create())
+
+}
+
+func TestResourceConnectionCreateAwsPrivateLink(t *testing.T) {
+	r := require.New(t)
+
+	b := newConnectionBuilder("privatelink_conn", "schema", "database")
+	b.ConnectionType("AWS PRIVATELINK")
+	b.PrivateLinkServiceName("com.amazonaws.us-east-1.materialize.example")
+	b.PrivateLinkAvailabilityZones([]string{"use1-az1", "use1-az2"})
+	r.Equal(`CREATE CONNECTION database.schema.privatelink_conn TO AWS PRIVATELINK (SERVICE NAME 'com.amazonaws.us-east-1.materialize.example',AVAILABILITY ZONES ('use1-az1', 'use1-az2'));`, b.Create())
+}
+
+func TestResourceConnectionCreateConfluentSchemaRegistry(t *testing.T) {
+	r := require.New(t)
+	b := newConnectionBuilder("csr_conn", "schema", "database")
+	b.ConnectionType("CONFLUENT SCHEMA REGISTRY")
+	b.ConfluentSchemaRegistryUrl("http://localhost:8081")
+	b.ConfluentSchemaRegistryUsername("user")
+	b.ConfluentSchemaRegistryPassword("password")
+	r.Equal(`CREATE CONNECTION database.schema.csr_conn TO CONFLUENT SCHEMA REGISTRY (URL 'http://localhost:8081', USERNAME = 'user', PASSWORD = SECRET password);`, b.Create())
+
 }
